fix: github checks has been reused
