+++ dashboard_dependencies_ymls/datafiles/django-splash_dependencies_health.yaml
@@ -1,9 +1,13 @@
-TIMESTAMP: 2023-11-30
+TIMESTAMP: 2023-12-01
 django:
     '4.0': 1.2.0
     '4.1': null
     '4.2': 1.3.0
     has_django: true
+github:
+    branch_count: 47
+    org_name: openedx
+    pulls_count: 2
 python:
     '3.10': null
     '3.11': null
